board: sunxi: fdt fixup of dram size for d1 platforms
On D1 boards, the U-Boot SPL detects the DRAM size and patches the DTB
accordingly. Later, if U-Boot uses a different DTB for booting Linux,
it copies only the dtb reserved-memory blocks, but not the dtb memory
block. Meanwhile, current mainline Linux device trees for D1 platforms
do not include a memory block. This change enables adding or updating
the memory block in the DTB passed to Linux.

Signed-off-by: Sergey Matyukevich <geomatsi@gmail.com>